# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Team Task Tracker API",
    "version": "1.0.0",
    "description": "REST API for team-based task management with JWT auth and RBAC"
  },
  "servers": [{ "url": "/api", "description": "API base path" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": { "type": "integer" },
          "code": { "type": "string" },
          "message": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register new organization and admin user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "name", "organizationName"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "name": { "type": "string" },
                  "organizationName": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "User registered" } }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Tokens issued" } }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Rotate refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": { "refreshToken": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "New tokens" } }
      }
    },
    "/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "List tasks with pagination and filters",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["TODO", "IN_PROGRESS", "IN_REVIEW", "DONE", "BLOCKED"] } },
          { "name": "priority", "in": "query", "schema": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH"] } },
          { "name": "assignee", "in": "query", "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Task list" } }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create task (Manager/Admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Task created" } }
      }
    },
    "/tasks/{id}/transition": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Transition task status",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["TODO", "IN_PROGRESS", "IN_REVIEW", "DONE", "BLOCKED"] }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Task updated" } }
      }
    },
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "User list" } }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create user (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "name", "role"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "name": { "type": "string" },
                  "role": { "type": "string", "enum": ["ADMIN", "MANAGER", "MEMBER"] }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "User created" } }
      }
    },
    "/users/{id}": {
      "patch": {
        "tags": ["Users"],
        "summary": "Update user (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "role": { "type": "string", "enum": ["ADMIN", "MANAGER", "MEMBER"] },
                  "password": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "User updated" } }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "204": { "description": "User deleted" } }
      }
    },
    "/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List projects",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Project list" } }
      }
    },
    "/analytics/overdue": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Overdue tasks per user and avg completion time",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Analytics data" } }
      }
    }
  }
}